+++ src/main/java/com/example/gym/service/TrainerService.java
@@ -24,7 +24,7 @@ public class TrainerService {
                 .collect(Collectors.toList());
     }
 
-    public TrainerDTO getTrainerById(Long id) {
+    public TrainerDTO getTrainerById(Integer id) {
         Trainer trainer = trainerRepository.findByIdAndIsActiveTrue(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Trainer not found with id: " + id));
         return modelMapper.map(trainer, TrainerDTO.class);
@@ -39,7 +39,7 @@ public class TrainerService {
     }
 
     @Transactional
-    public TrainerDTO updateTrainer(Long id, TrainerDTO trainerDTO) {
+    public TrainerDTO updateTrainer(Integer id, TrainerDTO trainerDTO) {
         Trainer existingTrainer = trainerRepository.findByIdAndIsActiveTrue(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Trainer not found with id: " + id));
 
@@ -49,7 +49,7 @@ public class TrainerService {
     }
 
     @Transactional
-    public void deactivateTrainer(Long id) {
+    public void deactivateTrainer(Integer id) {
         Trainer trainer = trainerRepository.findById(id)
                 .orElseThrow(() -> new ResourceNotFoundException("Trainer not found with id: " + id));
         trainer.setIsActive(false);
